+++ frontend/src/pages/integrations/alice/components/smart-device/components/device-skills/options-divider.tsx
@@ -1,0 +1,4 @@
+// Thin horizontal line between options in the popup
+export const OptionsDivider = () => (
+  <div className="aliceDeviceSkills-optionsDivider" />
+);
